+++ randovania/games/fusion/exporter/game_exporter.py
@@ -63,7 +63,9 @@ class FusionGameExporter(GameExporter[FusionGameExportParams]):
             if len(line) > 30:
                 text[index] = f"{line[0:27]}..."
 
-        patch_data["TitleText"] = [{"LineNum": index, "Text": line} for index, line in enumerate(text)]
+        patch_data["TitleText"] = [{"LineNum": index, "Text": line} for index, line in enumerate(text)] + patch_data[
+            "TitleText"
+        ]
 
         patcher.validate_patch_data(patch_data)
         try: